refactor: remove circular dependency on Main.ts
