+++ lana/src/AppSettings.ts
@@ -1,0 +1,5 @@
+/*
+ * Copyright (c) 2022 FinancialForce.com, inc. All rights reserved.
+ */
+
+export const appName: string = "Lana";
