+++ lana/src/Display.ts
@@ -3,12 +3,12 @@
  */
 
 import { ExtensionContext, Uri, window, workspace } from "vscode";
-import { appName } from "./Main";
+import { appName } from "./AppSettings";
 
 export class Display {
   private outputChannel = window.createOutputChannel(appName);
 
-  constructor(context: ExtensionContext) {}
+  constructor(context: ExtensionContext) { }
 
   output(message: string, showChannel: boolean = false) {
     if (showChannel) {
