+++ lana/src/Main.ts
@@ -10,8 +10,7 @@ interface InitConfig {
   debug: boolean;
 }
 
-export const appName: string = "Lana";
-export let context: Context | null = null;
+let context: Context | null = null;
 
 export function activate(
   extensionContext: ExtensionContext,
